+++ src/sim/single_set.cc
@@ -654,7 +654,7 @@ void SingleSet::uniAttach(FiberSet const& fibers)
 bool SingleSet::uniPrepare(PropertyList const& properties)
 {
     bool res = false;
-    size_t last = 0;
+    unsigned last = 0;
 
     PropertyList allprop = properties.find_all("single");
